mediamtx-camera-service/requirements.txt
# Core dependencies
websockets>=11.0
aiohttp>=3.8.0
PyYAML>=6.0
asyncio-mqtt>=0.11.0

# System monitoring
psutil>=5.9.0

# Camera discovery (optional - falls back to polling if not available)
pyudev>=0.24.0

# Security dependencies
PyJWT>=2.8.0
bcrypt>=4.0.0

# Development and testing (install with pip install -r requirements-dev.txt)
# pytest>=7.4.0
# pytest-asyncio>=0.21.1
# black>=23.7.0
# flake8>=6.0.0
# mypy>=1.5.0

# System Dependencies (install separately):
# Ubuntu/Debian: sudo apt-get install v4l-utils ffmpeg
# Required for camera capability detection and MediaMTX
